Refactor trace logging by introducing a new logging structure and updating filter application methods; remove obsolete formatter classes
