+++ lib/trace_viz/collectors/trace_point_collector.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "trace_viz/trace_data/trace_point_builder"
-require "trace_viz/trace_logger"
+require "trace_viz/loggers/trace_logger"
 require_relative "base_collector"
 require_relative "filters/registry"
 
@@ -11,21 +11,24 @@ module TraceViz
       def initialize(filters = [], sanitizers = [])
         super()
         @filters = Collectors::Filters::Registry.build(config.filters)
-        # @sanitizers = sanitizers
       end
 
       def collect(trace_point_event)
         trace_data = TraceData::TracePointBuilder.build(trace_point_event)
 
-        # Apply filters
-        return unless @filters.all? { |filter| filter.apply?(trace_data) }
+        return unless filters_apply?(trace_data)
 
-        # Apply sanitizers
-        # @sanitizers.each { |sanitizer| sanitizer.apply(trace_data) }
-
-        TraceLogger.log(trace_data)
+        Loggers::TraceLogger.log(trace_data)
 
         @collected_data << trace_data
+      end
+
+      private
+
+      attr_reader :filters
+
+      def filters_apply?(trace_data)
+        filters.all? { |filter| filter.apply?(trace_data) }
       end
     end
   end
