+++ lib/trace_viz/exporters/base_exporter.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module TraceViz
+  module Exporters
+    class BaseExporter
+      def initialize(collection)
+        @collection = collection
+      end
+
+      def export
+        raise NotImplementedError
+      end
+
+      private
+
+      attr_reader :collection
+    end
+  end
+end
