+++ lib/trace_viz/trace_data/trace_point/method_call.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "base"
-require "trace_viz/formatters/loggers/method_call_formatter"
 
 module TraceViz
   module TraceData
