+++ lib/trace_viz/trace_data/trace_point/method_return.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "base"
-require "trace_viz/formatters/loggers/method_return_formatter"
 
 module TraceViz
   module TraceData
